+++ R/transp_stomatalRegulationPlot.R
@@ -1,6 +1,6 @@
 transp_stomatalRegulationPlot<-function(x, soil, meteo, day, timestep, latitude, elevation, slope = NA, aspect = NA,
                                         type = "E") {
-  type = match.arg(type, c("E", "Ag","An" , "Gw", "T", "VPD"))
+  type = match.arg(type, c("E", "Ag","An" , "Gsw", "T", "VPD"))
   dctr = transp_transpirationSperry(x, soil, meteo, day, latitude, elevation, slope, aspect,
                                     stepFunctions = timestep, 
                                     modifyInputX = FALSE, modifyInputSoil = FALSE)
@@ -19,8 +19,8 @@ transp_stomatalRegulationPlot<-function(x, soil, meteo, day, timestep, latitude,
   Ag_shade = numeric(0)
   An_sunlit = numeric(0)
   An_shade = numeric(0)
-  Gw_sunlit = numeric(0)
-  Gw_shade = numeric(0)
+  Gsw_sunlit = numeric(0)
+  Gsw_shade = numeric(0)
   Temp_sunlit = numeric(0)
   Temp_shade = numeric(0)
   VPD_sunlit = numeric(0)
@@ -35,8 +35,8 @@ transp_stomatalRegulationPlot<-function(x, soil, meteo, day, timestep, latitude,
     Ag_shade = c(Ag_shade, phshade[[i]]$GrossPhotosynthesis)
     An_sunlit = c(An_sunlit, phsunlit[[i]]$NetPhotosynthesis)
     An_shade = c(An_shade, phshade[[i]]$NetPhotosynthesis)
-    Gw_sunlit = c(Gw_sunlit, phsunlit[[i]]$WaterVaporConductance)
-    Gw_shade = c(Gw_shade, phshade[[i]]$WaterVaporConductance)
+    Gsw_sunlit = c(Gsw_sunlit, phsunlit[[i]]$Gsw)
+    Gsw_shade = c(Gsw_shade, phshade[[i]]$Gsw)
     Temp_sunlit = c(Temp_sunlit, phsunlit[[i]]$LeafTemperature)
     Temp_shade = c(Temp_shade, phshade[[i]]$LeafTemperature)
     VPD_sunlit = c(VPD_sunlit, phsunlit[[i]]$LeafVPD)
@@ -50,13 +50,13 @@ transp_stomatalRegulationPlot<-function(x, soil, meteo, day, timestep, latitude,
     cohorts = c(cohorts, rep(cohnames[i], length(l[[i]]$psiLeaf)))
   }
   df_sunlit = data.frame(psi = psi, E = E, 
-                         Ag = Ag_sunlit, An = An_sunlit, Gw = Gw_sunlit,
+                         Ag = Ag_sunlit, An = An_sunlit, Gsw = Gsw_sunlit,
                          Temp = Temp_sunlit, VPD = VPD_sunlit, 
                          PM = PM_sunlit,
                          cohort = cohorts, leaf = "sunlit", 
                          stringsAsFactors = F)
   df_shade = data.frame(psi = psi, E = E, 
-                        Ag = Ag_shade, An = An_shade, Gw = Gw_shade,
+                        Ag = Ag_shade, An = An_shade, Gsw = Gsw_shade,
                         Temp = Temp_shade, VPD = VPD_shade, 
                         PM = PM_shade,
                         cohort = cohorts, leaf = "shade",
@@ -83,9 +83,9 @@ transp_stomatalRegulationPlot<-function(x, soil, meteo, day, timestep, latitude,
       geom_point(data = df_PM, aes_string(y="An", col = "cohort"))+
       ylab(expression(paste("Net photosynthesis "(mu*mol*C%.%s^{-1}%.%m^{-2}))))
   } 
-  else if(type=="Gw") {
-    g<- g + geom_path(aes_string(y = "Gw", col = "cohort"))+
-      geom_point(data = df_PM, aes_string(y="Gw", col = "cohort"))+
+  else if(type=="Gsw") {
+    g<- g + geom_path(aes_string(y = "Gsw", col = "cohort"))+
+      geom_point(data = df_PM, aes_string(y="Gsw", col = "cohort"))+
       ylab(expression(paste("Stomatal conductance "(mol%.%s^{-1}%.%m^{-2}))))
   } 
   else if(type=="T") {